pyproject.toml
[project]
name = "yeobi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.32.0",
    "openpyxl>=3.1.5",
    "pillow>=12.2.0",
    "pydantic>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
